+++ src/Facades/SP.php
@@ -3,11 +3,14 @@
 namespace Masterei\Sproc\Facades;
 
 use Illuminate\Support\Facades\Facade;
+use Masterei\Sproc\SP as BaseSP;
 
 class SP extends Facade
 {
     protected static function getFacadeAccessor()
     {
-        return \Masterei\Sproc\SP::class;
+        self::clearResolvedInstance(BaseSP::class);
+
+        return BaseSP::class;
     }
 }