+++ CHANGELOG.md
@@ -5,10 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2022-08-02
+
+### Fixed
+- Facade Bug: New Object Instance.
+- Resolved - self::clearResolvedInstance();
+
 ## [1.1.0] - 2022-07-18
 
 ### Added
-
 - hydrate() and toSql() method.
 - Please see [README File](README.md) for more information.